refactor: 파일 분리 & 로깅 추가
