+++ server-toss/src/modules/auth/constants/toss-api.constants.ts
@@ -1,0 +1,4 @@
+export const TOSS_API_ENDPOINTS = {
+  GENERATE_TOKEN: "/api-partner/v1/apps-in-toss/user/oauth2/generate-token",
+  LOGIN_ME: "/api-partner/v1/apps-in-toss/user/oauth2/login-me",
+} as const;
